Add page templates, route paths to templates
