+++ src/page-templates/topic-content-type/index.ts
@@ -1,0 +1,5 @@
+import TopicContentTypePage from './topic-content-type';
+import { TopicContentTypePageProps } from './types';
+
+export default TopicContentTypePage;
+export type { TopicContentTypePageProps };
