+++ src/page-templates/topic-content-type/topic-content-type.tsx
@@ -1,0 +1,16 @@
+import type { NextPage } from 'next';
+
+import { TopicContentTypePageProps } from './types';
+
+const TopicContentTypePage: NextPage<TopicContentTypePageProps> = ({
+    topicSlug,
+    contentTypeSlug,
+}) => {
+    return (
+        <div>
+            Topic: {topicSlug} | Content type: {contentTypeSlug}
+        </div>
+    );
+};
+
+export default TopicContentTypePage;
