+++ src/page-templates/topic-content-type/types.ts
@@ -1,0 +1,4 @@
+export interface TopicContentTypePageProps {
+    topicSlug: string;
+    contentTypeSlug: string;
+}
