+++ src/page-templates/topic-landing/index.ts
@@ -1,0 +1,5 @@
+import TopicLandingPage from './topic-landing';
+import { TopicLandingPageProps } from './types';
+
+export default TopicLandingPage;
+export type { TopicLandingPageProps };
