+++ src/page-templates/topic-landing/types.ts
@@ -1,0 +1,17 @@
+import { ITopicCard } from '../../components/topic-card/types';
+import { CTA } from '../../components/hero/types';
+import { ContentPiece } from '../../interfaces/content-piece';
+import { TertiaryNavItem } from '../../components/tertiary-nav/types';
+
+export interface TopicLandingPageProps {
+    name: string;
+    slug: string;
+    description: string;
+    ctas: CTA[];
+    topics: ITopicCard[];
+    relatedTopics: ITopicCard[];
+    featured: ContentPiece[];
+    content: ContentPiece[];
+    variant: 'light' | 'medium' | 'heavy';
+    tertiaryNavItems: TertiaryNavItem[];
+}
