+++ src/utils/maps.ts
@@ -1,0 +1,9 @@
+import { PillCategory, PillCategorySlugs } from '../types/pill-category';
+
+export const pillCategoryToSlug = new Map<PillCategory, PillCategorySlugs>([
+    ['Article', 'article'],
+    ['Demo App', 'demo-app'],
+    ['Tutorial', 'tutorial'],
+    ['Podcast', 'podcast'],
+    ['Video', 'video'],
+]);
